draw bone in props header
